Some slight improvements to selection of spells to cast.  (Fist versus Storm)
Consider the use of Harold's Bell for banishing gremlins.

A.M.C. gremlins never have Molybdenum tools.  Banish or defeat them without stasis.

Consolidate code for defeating gremlins that do not have Molybdenum tools.

When fighting a gremlin without a Mo tool, prefer using spells over attacking with one's weapon or using a dictionary.

Use Mafia's built-in tracking of opponents' HP, etc.

Add logic for "forceStasis" mode, in which we throw a fight in order to heal up in the underworld.

Preliminary code to use talismen of Renenutet against filthworms.

Add a couple of cases where we should not use a talisman of Renenutet, because it will most likely go to waste.

Better handling of astral bracers.  In particular, account for lowered spell MP costs.

Add a bunch of output for debugging and analysis of combat logic.