If we have lots of adventures, do allow non-turngen purchases, but ensure that the budget reflects that we intend to spend Ka on turngen as well.  (Probably doesn't affect purchases much unless we also have lots of Ka.)
Prioritize Renenutets over MP restorers.  This should ensure that we buy the talismen that are available on day one.

Don't ever keep more than a couple spirit beers on hand.

I've made an initial attempt to allow the script to be started from The Underworld, rather than the main map.  Not yet working code, because it attempts to use ed_edCombatHandler as a combat handler, yet that function resides in ed_combat.ash, which isn't imported into ed_edTheUndying.ash.  I might be able to just move the new ed_resume() function to ed_ascend.ash.  But since that change will span a couple files, I'll commit that change separately.