Update check that prevents wastage of talismen of Renenutet in short combats.
The check should now recognize (to within the current accuracy of our damage model for Storm of the Scarab) if we are able to defeat the opponent on the next round, using Storm of the Scarab.

Also, don't bother printing a message if we weren't about to use a talisman.