Cyrpt tweaks:
Don't do Cranny with +combat.

Don't do Alcove with +combat, either.  (Including Hippy Stench.)

Detect when we are near/at the Cranny or Alcove boss, and stop worrying about combat and ML modifiers when that happens.

Remove some non-Ed code.

Don't maintain excess turns of Blessing of Serqet, since we might be doing Snowmen next, in which case we will have to remove it or burn it doing something less desirable.