Improved use of talismen of Renenutet:
Generally, any time we would use Lash on an opponent should also be considered as an opportunity for using a talisman of Renenutet.  The logic for the two will share a function, ed_opponentHasDesiredItem, that detects most cases.  It will return true if the current opponent can drop an item from a generated list of desired items, and we currently do not have any of that item.  The only issue with this approach is that it will not handle obtaining more than one of an item.  I'll add special-case logic for Lash, and Renenutet will just ignore those, for now.

Some tweaks to forceStasis:  Allow stasis even when we only expect to have one more round available in the next stage.  Don't allow it, however, when our opponent might take many rounds to kill.  (i.e., Wisniewski)