Various changes:
Add support for non-muscle signs.

Ditch the non-compact charpane.  It's ugly and I can't find anything.  Use whatever you want; afaict, it doesn't affect the script.

Clarify the logic for adventuring at the Barrr.  Move LX_nastyBooty out to doTasks().

Add ed_ prefix to a couple settings.

Don't get Winklered until we've decided to do the nuns.

Get rid of redundant checks to see if we're Ed, as well as code that executes if we aren't.

Buy clovers in the main loop, rather than just at bedtime.

Try to avoid adventuring twice at our chosen semi-rare location, if for some reason we don't hit the counter auto-stop on the first.

Apply color to most of the print statements.