Various tweaks:
Tweaks for maintaining MP when low on meat at the start of the run.

Keep enough MP to banish pygmy orderlies.

There's no need to maintain Shelter of Shed in the Cranny once we are at the boss.

Support adventuring at the SMOOCH Army HQ at the start of the run.