Several changes:
Don't automatically maintain Blessing of Serqet at the first few adventuring locations, since we might not have Upgraded Legs.

Fix presumed typo in code that determines whether or not Purr of the Feline is beneficial with a Belly-Dancer.

Keep a bit more MP on hand when adventuring at the Secret Council Warehouse.