Various changes:
If "automation" preferences get changed while we have the user's previous settings stored, just clobber them.  Eventually, I expect the script will be good about always changing them back before it exits, so this case should be rare.

Add some settings in anticipation of SMOOCH Army HQ support.  These will automatically track when we should expect the boss fight.

Add a few maximization commands, to hopefully ensure that we don't go adventuring with sub-optimal equipment.  (Antique machetes and Talismen O' Nem, for example.)

Don't bother auto-selling stuff, as long as we have plenty of meat.

Don't buy the Hermit's clovers, until we have more than enough meat.

Set edDefeatAbort in a few key places.  Hopefully we won't ever see Mafia's auto-stop, but this at least puts it in line with how many undeaths we expect to use, so that when it does go off, it is likely that something unusual has happened.

Fix removal of mohawk wig after it gets equipped.

Buy and equip a third-hand lantern when we want initiative.  Every bit helps.

Remove some seemingly pointless buffs.

Removed the delay before fighting the Goblin King.

Adventuring at the Outskirts of the Knob during our first semi-rare window is now a separate task from "leg-day".

Added support for advancing the Lights Out quests.

Account for Muscle Percent modifiers, when preparing for The Horror.

Remove extra +10 margin-of-error when initially estimating damage from The Horror.

Added a few "handleMCD()" calls, in places where it might not have been turned up.

Maximize experience, when doing Black Forest, but include blackberry galoshes if we have them.

We now defer the F'c'le, if we have low skills and low maxhp.  Not being able to use a Renenutet there is very costly.

Added the cheap mus & mox buffs, when doing the F'c'le.  This is also in the hope of ensuring that we manage to use Renenutets.

Remove the separate loop for adventuring in the tavern basement.  We now return back to the top-level loop after each adventure.

Acquiring hermit clovers is lower priority than opening the Mysterious Island.

Adjust equipment every time through the main loop.

Using +stats-per-fight buffs is lower priority than consumption, in case we ever need to use adventures to craft our consumables.

I'm not sure why we have code that prevents us from adventuring at the Hidden Apartment Building when a semi-rare is coming up.  It should probably be removed.  For now, I just forced it to not skip out on us once we have Thrice-Cursed.  (Yeah, I'm sure it should just be removed.)

Bump up edDefeatAbort to 5 for Wisniewski and You the Adventurer.  This is more than enough for You.  On an unskilled run, Wisniewski can be even tougher, though.  This is fine for now.