ed_opponentHasDesiredItem changes:
Special-cases A-Boo clues, so that once we have three, we don't bother trying for more.

Add goat cheese!  And, special-cased to always want three.  (Although, in retrospect, it should instead desire cheese exactly when we have not finished that part of the quest...)

Experimenting with adding reodorant.  It may be beneficial to use a couple Lashes at the start of the run to get at least one.  Not sure if it's really worth a Renenutet.  For now, I consider this an experiment.