Various changes:
When we reach 4 deaths (1 Ka spent already) we always switch to "ed_edStatus" of "dying".

Remove apparently wasteful use of Curse of Indecision.  (Current code issues red warnings whenever the old logic would have used it.)

Don't bother using Renenutets to get tatters.  (We currently can't tell if Lash got one already, anyway.)

Account for today's not-yet-purchased Renenutets, when reserving them for the F'c'le.

When possible, use Curse of Indecision to try to buy time for a Renenutet.

Add some rudimentary code for detecting when our opponent is stunned.  This may allow us to get a Renenutet in.

During stasis, allow Fist and Mild Curse, if we are confident that they won't finish off our opponent.

If, for some reason, we don't even have Fist, fall back on attacking with our weapon.