Shopping tweaks, and some other minor changes:
Turn off Mafia's auto-recovery.  We'll handle it.

Collect an anticheese at the start of the run.

Keep up to 3 linen bandages.  Dunno why I had limited that to 1!

Allow purchasing of cure-alls.  But only if we have a fair amount of Ka.

Some tweaks to servant selection logic.  Always prefer Scribe to Priest if we have plenty of Ka.  Prefer Cat to both of them, if we are also level 13.

"Monster defeated in initialization" shouldn't get red text.  (In any of the four places where the message is displayed!)