Stasis tweaks:
Get rid of "test5" entirely, because it can trigger stasis at inappropriate times.

Don't stasis pirates after we have died a couple of times.  (Unless we are also flyering?)

If we somehow happen to be flyering, curse of fortune should prevent stasis.

If we are in danger of dying while curse of fortune is active, try to finish off the opponent.

Remove messages where the old code used to cast Curse of Indecision.

Remove messages when Talismen of Renenutet are kept in reserve against filthworms.