Tweaks to Ka-gathering:
Don't use Blessing of Serqet at the Hippy Camp unless jump_chance reports 100% without it.

Don't gather Ka at the Hippy Camp if we appear to be unable to handle the combats.  (Although, an exception is made so that our initial purchase of Upgraded Legs can happen there, since that's a reliable way to get to the Underworld.)

Allow Ka gathering at the SMOOCH Army HQ.