Use ed_predictMaximizationModifier to simplify logic to prepare for The Horror.  Also:
Make the check to see if we will cast Hide of Sobek slightly more accurate.  (Although, looking at it now, it's still wrong.)

Don't bother using buffs that we won't get in Ed Hardcore.

Do use a clover to get A-Boo clues, even if we think we have done so already.  (Because, maybe we didn't.)

And (not exactly related) don't print a message if we don't automatically create the Talisman O' Namsilat.  Try to create it, and only complain if that fails.