Various tweaks.  Importantly, when we want to shop, we now always set forceStasis.
Also, some allowances are made for inaccurate estimation of Storm of the Scarab damage, as well as damage dealt by some opponents.

Finishing a combat that is taking too long now takes priority (as it should) over fighting with just "attack with weapon".  This is important, because we currently do not check for physical resistance before attempting the latter.